Fix multithreading issues.
Also handle if a Session is Closed with an active transaction.
Rework a lot of the Commit/Abort code paths, and make use of the
Session.m object to properly handle Start/Abort/Commit.

Needs tests around Insert/Update/Upsert/Remove/Find and
Start/Commit/Abort sequencing.
