+++ session_test.go
@@ -4299,217 +4299,6 @@ func (s *S) TestVersionAtLeast(c *C) {
 	}
 }
 
-func (s *S) setup2Sessions(c *C) (*mgo.Session, *mgo.Collection, *mgo.Session, *mgo.Collection) {
-	// get the test infrastructure ready for doing transactions.
-	if !s.versionAtLeast(4, 0) {
-		c.Skip("transactions not supported before 4.0")
-	}
-	session1, err := mgo.Dial("localhost:40011")
-	c.Assert(err, IsNil)
-	// Collections must be created outside of a transaction
-	coll1 := session1.DB("mydb").C("mycoll")
-	err = coll1.Create(&mgo.CollectionInfo{})
-	session2 := session1.Copy()
-	if err != nil {
-		session1.Close()
-		c.Assert(err, IsNil)
-	}
-	coll2 := session2.DB("mydb").C("mycoll")
-	return session1, coll1, session2, coll2
-}
-func (s *S) TestTransactionInsertCommitted(c *C) {
-	session1, coll1, session2, coll2 := s.setup2Sessions(c)
-	defer session1.Close()
-	defer session2.Close()
-	c.Assert(session1.StartTransaction(), IsNil)
-	// call Abort in case there is a problem, but ignore an error if it was committed,
-	// otherwise the server will block in DropCollection because the transaction is active.
-	defer session1.AbortTransaction()
-	c.Assert(coll1.Insert(bson.M{"a": "a", "b": "b"}), IsNil)
-	var res bson.M
-	// Should be visible in the session that has the transaction
-	c.Assert(coll1.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "b"})
-	// Since the change was made in a transaction, session 2 should not see the document
-	err := coll2.Find(bson.M{"a": "a"}).One(&res)
-	c.Check(err, Equals, mgo.ErrNotFound)
-	c.Assert(session1.CommitTransaction(), IsNil)
-	// Now that it is committed, session2 should see it
-	err = coll2.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res)
-	c.Check(err, IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "b"})
-}
-
-func (s *S) TestTransactionInsertAborted(c *C) {
-	session1, coll1, session2, coll2 := s.setup2Sessions(c)
-	defer session1.Close()
-	defer session2.Close()
-	c.Assert(session1.StartTransaction(), IsNil)
-	// call Abort in case there is a problem, but ignore an error if it was committed,
-	// otherwise the server will block in DropCollection because the transaction is active.
-	defer session1.AbortTransaction()
-	c.Assert(coll1.Insert(bson.M{"a": "a", "b": "b"}), IsNil)
-	var res bson.M
-	// Should be visible in the session that has the transaction
-	c.Assert(coll1.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "b"})
-	// Since the change was made in a transaction, session 2 should not see the document
-	err := coll2.Find(bson.M{"a": "a"}).One(&res)
-	c.Check(err, Equals, mgo.ErrNotFound)
-	c.Assert(session1.AbortTransaction(), IsNil)
-	// Since it is Aborted, nobody should see the object
-	err = coll2.Find(bson.M{"a": "a"}).One(&res)
-	c.Check(err, Equals, mgo.ErrNotFound)
-	err = coll1.Find(bson.M{"a": "a"}).One(&res)
-	c.Check(err, Equals, mgo.ErrNotFound)
-
-}
-
-func (s *S) TestTransactionUpdateCommitted(c *C) {
-	session1, coll1, session2, coll2 := s.setup2Sessions(c)
-	defer session1.Close()
-	defer session2.Close()
-	c.Assert(coll1.Insert(bson.M{"a": "a", "b": "b"}), IsNil)
-	c.Assert(session1.StartTransaction(), IsNil)
-	// call Abort in case there is a problem, but ignore an error if it was committed,
-	// otherwise the server will block in DropCollection because the transaction is active.
-	defer session1.AbortTransaction()
-	c.Assert(coll1.Update(bson.M{"a": "a"}, bson.M{"$set": bson.M{"b": "c"}}), IsNil)
-	// Should be visible in the session that has the transaction
-	var res bson.M
-	c.Assert(coll1.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "c"})
-	// Since the change was made in a transaction, session 2 should not see it
-	c.Assert(coll2.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "b"})
-	c.Assert(session1.CommitTransaction(), IsNil)
-	// Now that it is committed, session2 should see it
-	c.Assert(coll2.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "c"})
-}
-
-func (s *S) TestTransactionUpdateAllCommitted(c *C) {
-	session1, coll1, session2, coll2 := s.setup2Sessions(c)
-	defer session1.Close()
-	defer session2.Close()
-	c.Assert(coll1.Insert(bson.M{"a": "a", "b": "b"}), IsNil)
-	c.Assert(coll1.Insert(bson.M{"a": "2", "b": "b"}), IsNil)
-	c.Assert(session1.StartTransaction(), IsNil)
-	// call Abort in case there is a problem, but ignore an error if it was committed,
-	// otherwise the server will block in DropCollection because the transaction is active.
-	defer session1.AbortTransaction()
-	changeInfo, err := coll1.UpdateAll(nil, bson.M{"$set": bson.M{"b": "c"}})
-	c.Assert(err, IsNil)
-	c.Check(changeInfo.Matched, Equals, 2)
-	c.Check(changeInfo.Updated, Equals, 2)
-	// Should be visible in the session that has the transaction
-	var res bson.M
-	c.Assert(coll1.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "c"})
-	c.Assert(coll1.Find(bson.M{"a": "2"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "2", "b": "c"})
-	// Since the change was made in a transaction, session 2 should not see it
-	c.Assert(coll2.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "b"})
-	c.Assert(coll2.Find(bson.M{"a": "2"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "2", "b": "b"})
-	c.Assert(session1.CommitTransaction(), IsNil)
-	// Now that it is committed, session2 should see it
-	c.Assert(coll2.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "c"})
-	c.Assert(coll2.Find(bson.M{"a": "2"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "2", "b": "c"})
-}
-
-func (s *S) TestTransactionUpsertCommitted(c *C) {
-	session1, coll1, session2, coll2 := s.setup2Sessions(c)
-	defer session1.Close()
-	defer session2.Close()
-	c.Assert(coll1.Insert(bson.M{"a": "a", "b": "b"}), IsNil)
-	c.Assert(session1.StartTransaction(), IsNil)
-	// call Abort in case there is a problem, but ignore an error if it was committed,
-	// otherwise the server will block in DropCollection because the transaction is active.
-	defer session1.AbortTransaction()
-	// One Upsert updates, the other Upsert creates
-	changeInfo, err := coll1.Upsert(bson.M{"a": "a"}, bson.M{"$set": bson.M{"b": "c"}})
-	c.Assert(err, IsNil)
-	c.Check(changeInfo.Matched, Equals, 1)
-	c.Check(changeInfo.Updated, Equals, 1)
-	changeInfo, err = coll1.Upsert(bson.M{"a": "2"}, bson.M{"$set": bson.M{"b": "c"}})
-	c.Assert(err, IsNil)
-	c.Check(changeInfo.Matched, Equals, 0)
-	c.Check(changeInfo.UpsertedId, NotNil)
-	// Should be visible in the session that has the transaction
-	var res bson.M
-	c.Assert(coll1.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "c"})
-	c.Assert(coll1.Find(bson.M{"a": "2"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "2", "b": "c"})
-	// Since the change was made in a transaction, session 2 should not see it
-	c.Assert(coll2.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "b"})
-	c.Assert(coll2.Find(bson.M{"a": "2"}).One(&res), Equals, mgo.ErrNotFound)
-	c.Assert(session1.CommitTransaction(), IsNil)
-	// Now that it is committed, session2 should see it
-	c.Assert(coll2.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "c"})
-	c.Assert(coll2.Find(bson.M{"a": "2"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "2", "b": "c"})
-}
-
-func (s *S) TestTransactionRemoveCommitted(c *C) {
-	session1, coll1, session2, coll2 := s.setup2Sessions(c)
-	defer session1.Close()
-	defer session2.Close()
-	c.Assert(coll1.Insert(bson.M{"a": "a", "b": "b"}), IsNil)
-	c.Assert(session1.StartTransaction(), IsNil)
-	// call Abort in case there is a problem, but ignore an error if it was committed,
-	// otherwise the server will block in DropCollection because the transaction is active.
-	defer session1.AbortTransaction()
-	c.Assert(coll1.Remove(bson.M{"a": "a"}), IsNil)
-	// Should be gone in the session that has the transaction
-	var res bson.M
-	c.Assert(coll1.Find(bson.M{"a": "a"}).One(&res), Equals, mgo.ErrNotFound)
-	// Since the change was made in a transaction, session 2 should still see the document
-	c.Assert(coll2.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "b"})
-	c.Assert(session1.CommitTransaction(), IsNil)
-	// Now that it is committed, it should be gone
-	c.Assert(coll1.Find(bson.M{"a": "a"}).One(&res), Equals, mgo.ErrNotFound)
-	c.Assert(coll2.Find(bson.M{"a": "a"}).One(&res), Equals, mgo.ErrNotFound)
-}
-
-func (s *S) TestTransactionRemoveAllCommitted(c *C) {
-	session1, coll1, session2, coll2 := s.setup2Sessions(c)
-	defer session1.Close()
-	defer session2.Close()
-	c.Assert(coll1.Insert(bson.M{"a": "a", "b": "b"}), IsNil)
-	c.Assert(coll1.Insert(bson.M{"a": "2", "b": "b"}), IsNil)
-	c.Assert(session1.StartTransaction(), IsNil)
-	// call Abort in case there is a problem, but ignore an error if it was committed,
-	// otherwise the server will block in DropCollection because the transaction is active.
-	defer session1.AbortTransaction()
-	changeInfo, err := coll1.RemoveAll(bson.M{"a": bson.M{"$exists": true}})
-	c.Assert(err, IsNil)
-	c.Check(changeInfo.Matched, Equals, 2)
-	c.Check(changeInfo.Removed, Equals, 2)
-	// Should be gone in the session that has the transaction
-	var res bson.M
-	c.Assert(coll1.Find(bson.M{"a": "a"}).One(&res), Equals, mgo.ErrNotFound)
-	c.Assert(coll1.Find(bson.M{"a": "2"}).One(&res), Equals, mgo.ErrNotFound)
-	// Since the change was made in a transaction, session 2 should still see the document
-	c.Assert(coll2.Find(bson.M{"a": "a"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "a", "b": "b"})
-	c.Assert(coll2.Find(bson.M{"a": "2"}).Select(bson.M{"a": 1, "b": 1, "_id": 0}).One(&res), IsNil)
-	c.Check(res, DeepEquals, bson.M{"a": "2", "b": "b"})
-	c.Assert(session1.CommitTransaction(), IsNil)
-	// Now that it is committed, it should be gone
-	c.Assert(coll1.Find(bson.M{"a": "a"}).One(&res), Equals, mgo.ErrNotFound)
-	c.Assert(coll1.Find(bson.M{"a": "2"}).One(&res), Equals, mgo.ErrNotFound)
-	c.Assert(coll2.Find(bson.M{"a": "a"}).One(&res), Equals, mgo.ErrNotFound)
-	c.Assert(coll2.Find(bson.M{"a": "2"}).One(&res), Equals, mgo.ErrNotFound)
-}
-
 // --------------------------------------------------------------------------
 // Some benchmarks that require a running database.
